Remove compiler warnings in GLR parsers.
	* data/glr.c (yySymbol): Define as int to avoid compiler warnings
	about possible change of value.

+++ data/glr.c
@@ -673,7 +673,7 @@ typedef int yyStateNum;
 typedef int yyRuleNum;
 
 /** Grammar symbol */
-typedef short int yySymbol;
+typedef int yySymbol;
 
 /** Item references, as in LALR(1) machine */
 typedef short int yyItemNum;
